+++ contracts/helpers/TestLiquidityProtection.sol
@@ -12,7 +12,6 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
     uint256 private _poolTokenRateD;
 
     constructor(
-        IBancorNetworkV3 networkV3,
         address payable vaultV3,
         ILiquidityProtectionSettings settings,
         ILiquidityProtectionStore store,
@@ -24,7 +23,6 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
     )
         public
         LiquidityProtection(
-            networkV3,
             vaultV3,
             settings,
             store,
@@ -63,17 +61,13 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
         return (impLossRate.n, impLossRate.d);
     }
 
-    function compensationAmountTest(
-        uint256 amount,
+    function deductILTest(
         uint256 total,
         uint256 lossN,
-        uint256 lossD,
-        uint256 levelN,
-        uint256 levelD
+        uint256 lossD
     ) external pure returns (uint256) {
         Fraction memory loss = Fraction({ n: lossN, d: lossD });
-        Fraction memory level = Fraction({ n: levelN, d: levelD });
-        return _compensationAmount(amount, total, loss, level);
+        return _deductIL(total, loss);
     }
 
     function averageRateTest(IDSToken poolToken, IReserveToken reserveToken) external view returns (uint256, uint256) {
@@ -83,6 +77,8 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
     }
 
     function removeLiquidityTargetAmountTest(
+        IDSToken poolToken,
+        IReserveToken reserveToken,
         uint256 poolTokenRateN,
         uint256 poolTokenRateD,
         uint256 poolAmount,
@@ -92,9 +88,7 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
         uint128 removeSpotRateN,
         uint128 removeSpotRateD,
         uint128 removeAverageRateN,
-        uint128 removeAverageRateD,
-        uint256 addTimestamp,
-        uint256 removeTimestamp
+        uint128 removeAverageRateD
     ) external returns (uint256) {
         _poolTokenRateOverride = true;
         _poolTokenRateN = poolTokenRateN;
@@ -109,14 +103,12 @@ contract TestLiquidityProtection is LiquidityProtection, TestTime {
             removeAverageRateD: removeAverageRateD
         });
 
-        uint256 targetAmount = _removeLiquidityTargetAmount(
-            IDSToken(0),
-            IReserveToken(0),
+        (uint256 targetAmount,) = _removeLiquidityAmounts(
+            poolToken,
+            reserveToken,
             poolAmount,
             reserveAmount,
-            packedRates,
-            addTimestamp,
-            removeTimestamp
+            packedRates
         );
         _poolTokenRateOverride = false;
         return targetAmount;